% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jonesDatum.R
\name{equalizeHist}
\alias{equalizeHist}
\title{equalizeHist}
\usage{
equalizeHist(imgMat, bitDepth = 2^8, ...)
}
\arguments{
\item{imgMat}{matrix, must be 2D}

\item{bitDepth}{int, number of bins for the hist}

\item{...}{unsused}
}
\description{
equalizeHist
}
