% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fileImporter.R
\name{fileImporter}
\alias{fileImporter}
\title{fileImporter: Load Excel Data From Jones Group}
\usage{
fileImporter(filePath, dataPath = NULL, verbose = FALSE,
  dataPathAppend = "", ...)
}
\arguments{
\item{filePath}{character: full path to excel file (required)}

\item{dataPath}{character: full path to data file for appending new data (optional)}

\item{verbose}{bool: Report progress of file loading, default FALSE (recommended TRUE)}

\item{dataPathAppend}{chracter: string to append onto the saved datafile, used only if \code{dataPath} is supplied}

\item{...}{Arguments passed onto internal functions.}
}
\value{
List with the following Structure. \cr
  \describe{
   \item{Background}{List that contains, if applicable, the background lum (0-255), the ROI area and ROI coordinates.
   \describe{
     \item{ID}{Which ROI ID was used for extraction of the background data.}
     \item{Data}{The Time series calcium lum (0-255) of the background roi.}
     \item{ROI }{Coordinates of the ROI in X, Y, X.Pixels and Y.Pixels.}
     \item{Area}{The Area in microns squared.}
   }}
   \item{Areas}{data.frame with cell id and area of ROI in microns squared.}
   \item{ROIs}{list of ROI coordinates (named) for each cellID}
   \describe{
     \item{CellID}{Named list of data.frame\cr
     \describe{
       \item{X}{}
       \item{Y}{ }
       \item{X.Pixels}{ }
       \item{Y.Pixels}{ }
     }}
   }
   \item{Raw}{ }
   \item{File}{
   \describe{
     \item{Name}{ }
     \item{Properties}{matrix of values from excel header.}
   }}
   \item{Info}{
   \describe{
     \item{SamplingRate}{ }
     \item{TreatmentStart}{ }
     \item{nCells}{ }
     \item{HasBackground}{ }
   }}
   \item{Original}{data.frame of CellID without Time, but includes bg if applicable.}
   \item{Time}{ }
   \item{Data}{df/f0}
   \item{Corrected}{df/f0 - background}
   \item{Standardized}{((data-median)/mad)-mean}
   \item{Image}{
   \describe{
     \item{HasImage}{ }
     \item{File}{
     \describe{
       \item{Name}{ }
       \item{Type}{ }
       \item{Path}{ }
     }}
     \item{Data}{rtiff::readTiff}
   }}
  }
}
\description{
fileImporter: Load Excel Data From Jones Group
}
