% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jonesDatum.R
\name{fitEllipse}
\alias{fitEllipse}
\title{fitEllipse}
\usage{
fitEllipse(x, y = NULL)
}
\arguments{
\item{x}{vector of Xs or data frame with 2 columns}

\item{y}{vector of Ys or NOT}
}
\description{
fitEllipse
}
