% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fileImporter.R
\name{inspector}
\alias{inspector}
\title{inspector}
\usage{
inspector(cur.file, col.drop = NULL, do.View = TRUE, all.sheets = TRUE,
  main.sheet = 1, ROI.scale = 1, verbose = FALSE, ...)
}
\arguments{
\item{cur.file}{File name for file being loaded}

\item{col.drop}{Name of a column to drop, typically 'Frame'}

\item{do.View}{Causes a halt in execution to view the data in a window}

\item{all.sheets}{Load all the sheets or just the main?}

\item{main.sheet}{Index of main sheet, 1 for first.}

\item{ROI.scale}{A scale factor for converting the xy positions to pixels on image}

\item{verbose}{Chatty, much?}

\item{...}{Passed on to other methods}
}
\description{
inspector
}
