% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jonesDatum.R
\docType{class}
\name{jonesDatum-class}
\alias{jonesDatum-class}
\alias{jonesDatum}
\alias{plot,jonesDatum,character-method}
\alias{subset,jonesDatum-method}
\title{Jones Datum object}
\usage{
\S4method{plot}{jonesDatum,character}(x, y = c("raw", "adjusted",
  "standardized", "background", "corrected"), xlab = "", ylab = "",
  axes = FALSE, ...)

\S4method{subset}{jonesDatum}(x, subset, select, type = c("adjusted", "raw",
  "standardized", "background", "corrected"), drop = FALSE, ...)
}
\arguments{
\item{x}{jonesDatum object, required}

\item{y}{char of: 'raw', 'adjusted', 'scaled', 'background', 'corrected'}

\item{xlab}{X-axis label (units will be appended)}

\item{ylab}{Y-axis label (units will be appended)}

\item{axes}{bool, draw with or without axes (FALSE by default)}

\item{...}{for plot: passed to plot.window. otherwise unused.}

\item{subset}{logical test for rows}

\item{select}{logical test for columns}

\item{type}{type=c('adjusted', 'standardized','raw','background','corrected')}

\item{drop}{sent to indexing operation for \code{x[,,drop]}}
}
\description{
Jones Datum object

plot

subset
}
\section{Fields}{

\describe{
\item{\code{fileName}}{Provide the full path to teh file}

\item{\code{metaData}}{no input. contains metadata and info for file}

\item{\code{data}}{no input. list of data types: Raw, Adjusted, Rescaled and Time vector}

\item{\code{background}}{no input. contains, if applicable, lum, roi and area of background roi}

\item{\code{ares}}{no input. data.frame of cell# and corresponding roi area}

\item{\code{rois}}{no input. named list for each roi in the dataset}

\item{\code{image}}{no input. Contains the image data, if applicable.}
}}

\section{Methods}{

\describe{
\item{\code{getConfluence()}}{Report confluence as a percent (0-100).}

\item{\code{getData(type = c("adjusted", "standardized", "raw", "background",
  "corrected"))}}{Extract full dataset of provided type.}

\item{\code{getDataAsList(type = c("adjusted", "standardized", "raw", "background",
  "corected"))}}{Extract the data as a named list of X,Y datasets for each Cell.}

\item{\code{getIDs()}}{Get the names of the cell IDs}

\item{\code{getOtherTreatment()}}{Check the other treatments, typically IL-1b.}

\item{\code{getTreatment()}}{Report Treatment type, typically ET1.}

\item{\code{isConfluent(prcnt = 100)}}{Check if confluence is == \code{prcnt}.}

\item{\code{isET1()}}{Check if ET1 is in the primary treatment.}

\item{\code{isIL1b()}}{Check if IL1b is in other treatments.}

\item{\code{plotImage(channel = c("green", "all", "red", "blue"), ofst = 0,
  greyscale = FALSE, equalize = TRUE, gamma = 2.2, colors = NULL,
  texts = NULL, draw.roi = TRUE, roi.alpha = 0.5, draw.text = TRUE,
  text.color = NULL, ...)}}{Plots the corresponding image if applicable. \cr
\code{channel}:    Choose a specific channel or 'all' for all channels.\cr
\code{ofst}:       Value to set unchosen channels on [0,1] can reduce contrast\cr
\code{greyscale}:  boolean, convert image to greyscale (2d indexed)\cr
\code{equalize}:   boolean, Use histogram equalization\cr
\code{gamma}:      double, used if greyscale or hist equalization\cr
\code{colors}:     hex color matrix for ROIs (defaults to all red)\cr
\code{draw.roi}    boolean, Draw the ROIs over the image \cr
\code{roi.alpha}   double, Transparency level for ROI on [0,1]\cr
\code{draw.text}   boolean, Draw text labels, provided or otheriwse, at the centroid of each ROI\cr
\code{text.color}  array, Hex color vector. If scalar, color is applied to all text, otherwise
it must be the same length as .self$nCells().\cr
\code{...}:        Parameters passed to \code{plot.window()}.}

\item{\code{save(pathName, objectName = "", ...)}}{Save the current object on the file in R external object format.}
}}

