% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fileImporter.R
\name{locateData}
\alias{locateData}
\title{locateData}
\usage{
locateData(fileName, altName = NULL, extension = "rda")
}
\arguments{
\item{fileName}{data file path/name}

\item{altName}{possible optional name to search for if fileName isn't found}

\item{extension}{Data file extension}
}
\description{
Recursively look through directories for a data file
}
