% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fileImporter.R
\name{locateImage}
\alias{locateImage}
\title{locateImage}
\usage{
locateImage(baseName, parent, extension = "tif")
}
\arguments{
\item{baseName}{File's base name}

\item{parent}{Parent directory to search in for baseName}

\item{extension}{Image file extension}
}
\description{
Locate the image file associated with an excel data file.
}
