% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fileImporter.R
\name{processSheets}
\alias{processSheets}
\title{processSheets}
\usage{
processSheets(cur.File, sheets, ROI.scale = 1, verbose = FALSE, ...)
}
\arguments{
\item{cur.File}{File name for file being loaded}

\item{sheets}{names of sheets to process}

\item{ROI.scale}{A scale factor for converting the xy positions to pixels on image}

\item{verbose}{Chatty, much?}

\item{...}{Passed on to other methods}
}
\description{
Collect the data from the excel sheets
}
